requirements.txt
adapter-transformers~=3.0.1
torch~=1.12.0
pandas~=1.4.3
numpy~=1.23.1
nltk~=3.7
tqdm~=4.64.0
datasets~=2.3.2
torchtext~=0.13.0
rich~=12.5.1
spacy~=3.4.0
setuptools~=60.2.0
sentence-transformers~=2.2.2
spacy-alignments~=0.8.5